Huge corrections regarding dependencies between resources
After several tests, it seems there are many problems with dependencies,
especially when we're wanting to get the whole stuff set in one row.

The "exec dovecot" resource wasn't of any use, really. In additions,
some files were to be created *before* the package is installed, and
this made puppet crash with "no such file or directory" error.

The modifications allow to get a smooth, one-shot installation, with the
whole dependencies in the right order:
- packages
- then configurations
- then service